+++ fabric/src/main/java/dev/xylonity/companions/platform/CompanionsFabricPlatform.java
@@ -1,5 +1,10 @@
 package dev.xylonity.companions.platform;
 
 public class CompanionsFabricPlatform implements CompanionsPlatform {
-    ;;
+
+    @Override
+    public boolean requiresGlobalTeslaRenderer() {
+        return true;
+    }
+
 }